Issue #561: ICU date difference
Factor out some common utilities for reuse.
