+++ extension/icu/CMakeLists.txt
@@ -6,7 +6,7 @@ include_directories(include)
 
 add_library(
   icu_extension STATIC icu-collate.cpp icu-extension.cpp icu-dateadd.cpp
-                       icu-datepart.cpp icu-datetrunc.cpp)
+                       icu-datefunc.cpp icu-datepart.cpp icu-datetrunc.cpp)
 link_threads(icu_extension)
 disable_target_warnings(icu_extension)
